Add files via upload
The above function file is to find if the (year,month,date) entered are correct according to the calender or not. I eliminated the major hurdle of eliminating 1900th year as non leap year whereas 2000th year is a leap year.
Learning outcome: Use of AND, OR and NOT operator in relation with correct input of the paranthesis.